rename ProductItemsBadge to BadgeOverlay
