+++ src/components/BadgeOverlay/BadgeOverlay.test.js
