+++ src/components/BadgeOverlay/styles.js
